requirements.txt
# Core Framework
Django==4.2.16
djangorestframework==3.15.2

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.1

# CORS
django-cors-headers==4.4.0

# Environment Variables
python-dotenv==1.0.1

# Production server
gunicorn==22.0.0

# Development & Testing
pytest==8.3.3
pytest-django==4.9.0
factory-boy==3.3.1
faker==30.3.0
